+++ Examples/FluxClientSimpleExample.cs
@@ -11,7 +11,7 @@ namespace Examples
             Console.WriteLine("Start");
 
             var options = new FluxConnectionOptions("http://127.0.0.1:8086", TimeSpan.FromSeconds(20));
-            var client = FluxClientFactory.Create(options);
+            using var client = FluxClientFactory.Create(options);
 
             var fluxQuery = "from(bucket: \"telegraf\")\n"
                                + " |> filter(fn: (r) => (r[\"_measurement\"] == \"cpu\" AND r[\"_field\"] == \"usage_system\"))"